requirements.txt
aiohttp
sseclient-py